move Spinner into sharedComponents
