+++ apps/src/aiTutor/views/teacherDashboard/AccessControls.tsx
@@ -4,7 +4,7 @@ import * as Table from 'reactabular-table';
 
 import {fetchStudents} from '@cdo/apps/aiTutor/accessControlsApi';
 import {StudentAccessData} from '@cdo/apps/aiTutor/types';
-import Spinner from '@cdo/apps/code-studio/pd/components/spinner';
+import Spinner from '@cdo/apps/sharedComponents/Spinner';
 import {tableLayoutStyles as tableStyles} from '@cdo/apps/templates/tables/tableConstants';
 
 import {styleOverrides} from './InteractionsTable';
