+++ apps/src/code-studio/assets/DefaultSpritesEditor.jsx
@@ -5,8 +5,8 @@ import {
   regenerateDefaultSpriteMetadata,
 } from '@cdo/apps/assetManagement/animationLibraryApi';
 import DefaultSpriteRow from '@cdo/apps/code-studio/assets/DefaultSpriteRow';
-import Spinner from '@cdo/apps/code-studio/pd/components/spinner';
 import Button from '@cdo/apps/legacySharedComponents/Button';
+import Spinner from '@cdo/apps/sharedComponents/spinner';
 
 export default class DefaultSpritesEditor extends React.Component {
   state = {
