+++ apps/src/code-studio/components/libraries/ShareTeacherLibraries.jsx
@@ -3,8 +3,8 @@ import React from 'react';
 import {connect} from 'react-redux';
 
 import SortedTableSelect from '@cdo/apps/code-studio/components/SortedTableSelect';
-import Spinner from '@cdo/apps/code-studio/pd/components/spinner';
 import Button from '@cdo/apps/legacySharedComponents/Button';
+import Spinner from '@cdo/apps/sharedComponents/Spinner';
 import {
   setPersonalProjects,
   updateProjectLibrary,
