+++ apps/src/code-studio/components/pairing/Pairing.jsx
@@ -2,8 +2,8 @@ import $ from 'jquery';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import Spinner from '@cdo/apps/code-studio/pd/components/spinner';
 import firehoseClient from '@cdo/apps/lib/util/firehose';
+import Spinner from '@cdo/apps/sharedComponents/Spinner.jsx';
 import color from '@cdo/apps/util/color';
 import i18n from '@cdo/locale';
