+++ apps/src/code-studio/pd/foorm/editor/components/FoormEntityEditorHeader.jsx
@@ -7,7 +7,7 @@ import {connect} from 'react-redux';
 import FontAwesome from '@cdo/apps/legacySharedComponents/FontAwesome';
 import ToggleGroup from '@cdo/apps/templates/ToggleGroup';
 
-import Spinner from '../../../components/spinner';
+import Spinner from '../../../../../sharedComponents/Spinner';
 
 const PREVIEW_ON = 'preview-on';
 const PREVIEW_OFF = 'preview-off';
