+++ apps/src/code-studio/pd/form_components_func/FormController.jsx
@@ -5,9 +5,9 @@ import PropTypes from 'prop-types';
 import React, {useState, useEffect, useCallback, useMemo} from 'react';
 import {Button, Alert, FormGroup} from 'react-bootstrap'; // eslint-disable-line no-restricted-imports
 
-import Spinner from '@cdo/apps/code-studio/pd/components/spinner';
 import {EVENTS} from '@cdo/apps/lib/util/AnalyticsConstants';
 import analyticsReporter from '@cdo/apps/lib/util/AnalyticsReporter';
+import Spinner from '@cdo/apps/sharedComponents/Spinner';
 import usePrevious from '@cdo/apps/util/usePrevious';
 import i18n from '@cdo/locale';
