+++ apps/src/code-studio/pd/workshop_dashboard/attendance/workshop_attendance.jsx
@@ -11,7 +11,7 @@ import {connect} from 'react-redux';
 
 import color from '@cdo/apps/util/color';
 
-import Spinner from '../../components/spinner';
+import Spinner from '../../../../sharedComponents/Spinner';
 import SessionTime from '../components/session_time';
 import {
   PermissionPropType,
