+++ apps/src/code-studio/pd/workshop_dashboard/components/workshop_table_loader.jsx
@@ -11,7 +11,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import Spinner from '../../components/spinner';
+import Spinner from '../../../../sharedComponents/Spinner';
 
 export default class WorkshopTableLoader extends React.Component {
   static propTypes = {
