+++ apps/src/code-studio/pd/workshop_dashboard/legacy_survey_summaries.jsx
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import Spinner from '../components/spinner';
+import Spinner from '../../../sharedComponents/Spinner';
 
 class LegacySurveySummaryTable extends React.Component {
   static propTypes = {
