+++ apps/src/code-studio/pd/workshop_dashboard/reports/local_summer_workshop_daily_survey/results_loader.jsx
@@ -4,7 +4,7 @@ import React from 'react';
 
 import color from '@cdo/apps/util/color';
 
-import Spinner from '../../../components/spinner';
+import Spinner from '../../../../../sharedComponents/Spinner';
 
 import Results from './results';
