+++ apps/src/code-studio/pd/workshop_dashboard/reports/teacher_attendance_report.jsx
@@ -6,7 +6,7 @@ import React from 'react';
 import {Button} from 'react-bootstrap'; // eslint-disable-line no-restricted-imports
 import {connect} from 'react-redux';
 
-import Spinner from '../../components/spinner';
+import Spinner from '../../../../sharedComponents/Spinner';
 import {PermissionPropType} from '../permission';
 
 import {QUERY_BY_VALUES, COURSE_VALUES} from './report_constants';
