+++ apps/src/code-studio/pd/workshop_dashboard/workshop.jsx
@@ -11,7 +11,7 @@ import React from 'react';
 import {Grid, Row, Col} from 'react-bootstrap'; // eslint-disable-line no-restricted-imports
 import {connect} from 'react-redux';
 
-import Spinner from '../components/spinner';
+import Spinner from '../../../sharedComponents/Spinner';
 
 import AttendancePanel from './AttendancePanel';
 import DetailsPanel from './DetailsPanel';
