+++ apps/src/music/views/SaveStatus.tsx
@@ -2,8 +2,8 @@ import classNames from 'classnames';
 import React from 'react';
 import {useSelector} from 'react-redux';
 
-import Spinner from '@cdo/apps/code-studio/pd/components/spinner';
 import {projectUpdatedStatuses} from '@cdo/apps/code-studio/projectRedux';
+import Spinner from '@cdo/apps/sharedComponents/Spinner';
 import {commonI18n} from '@cdo/apps/types/locale';
 
 const moduleStyles = require('./save-status.module.scss').default;
