+++ apps/src/templates/LoadableComponent.jsx
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React, {useEffect, useState} from 'react';
 
-import Spinner from '@cdo/apps/code-studio/pd/components/spinner';
+import Spinner from '@cdo/apps/sharedComponents/Spinner';
 import color from '@cdo/apps/util/color';
 import i18n from '@cdo/locale';
