+++ apps/src/templates/instructions/DocumentationTab.jsx
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import React, {useState, useEffect, forwardRef} from 'react';
 import {connect} from 'react-redux';
 
-import Spinner from '@cdo/apps/code-studio/pd/components/spinner';
 import FontAwesome from '@cdo/apps/legacySharedComponents/FontAwesome';
+import Spinner from '@cdo/apps/sharedComponents/Spinner';
 import {useFetch} from '@cdo/apps/util/useFetch';
 import i18n from '@cdo/locale';
