+++ apps/src/templates/instructions/codeReviewV2/Comment.jsx
@@ -3,10 +3,10 @@ import PropTypes from 'prop-types';
 import React, {useState, useEffect} from 'react';
 import {connect} from 'react-redux';
 
-import Spinner from '@cdo/apps/code-studio/pd/components/spinner';
 import {ViewType} from '@cdo/apps/code-studio/viewAsRedux';
 import fontConstants from '@cdo/apps/fontConstants';
 import FontAwesome from '@cdo/apps/legacySharedComponents/FontAwesome';
+import Spinner from '@cdo/apps/sharedComponents/Spinner';
 import InlineDropdownMenu from '@cdo/apps/templates/InlineDropdownMenu';
 import {reviewCommentShape} from '@cdo/apps/templates/instructions/codeReviewV2/shapes';
 import SafeMarkdown from '@cdo/apps/templates/SafeMarkdown';
